+++ src/lib/html.ts
@@ -64,7 +64,7 @@ function test(parser: (html: string) => Document): boolean {
 `;
     const doc = parser(html);
     switch (false) {
-      case doc.URL && decodeURI(doc.URL) === decodeURI(window.location.href):
+      case doc.URL && doc.URL.startsWith(`${window.location.protocol}//${window.location.host}`):
       case doc.title === '&':
       case !!doc.querySelector('html.html[lang="en"]'):
       case !!doc.querySelector('head>link')!['href']: